fix the following warning on FreeBSD: {{{ ./config_openssl.c:23: warning: return makes integer from pointer without a cast }}} FreeBSD's pthread_t is a pointer type, but this cast to unsigned long is safe, because all UNIX variants use ILP32 or LP64 (thus, unsigned long can hold a pointer). Althouth this is not safe on 64bit Windows (LLP64).
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.6@10018 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ lib/libgfarm/gfarm/config_openssl.c
@@ -20,7 +20,7 @@ static unsigned long
 gfarm_openssl_threadid(void)
 {
 	/* XXX - check pthread_self() returns unsigned long or not */
-	return (pthread_self());
+	return ((unsigned long)pthread_self());
 }
 
 /*
